feat: add Float32 and Float64 types, default Float to 32-bit
- Added Float32 and Float64 to the type system.
- Updated Float to be 32-bit by default (previously 64-bit).
- Enhanced Binder to handle floating-point promotion (Float32 -> Float64).
- Implemented SIToFP/UIToFP, FPToSI/FPToUI, and FPExt/FPTrunc in LLVM codegen.
- Updated stdlib with printFloat (32-bit) and printFloat64 (64-bit).
- Updated ConstFloat to use double for literal precision.

+++ src/ast/const/const_float.hpp
@@ -3,9 +3,9 @@
 
 class ConstFloat : public ExpressionNode {
 public:
-  float value_;
+  double value_;
   ConstFloat() noexcept = default;
-  ConstFloat(float value) noexcept : value_(value) {}
+  ConstFloat(double value) noexcept : value_(value) {}
 
   void accept(Visitor &v) override { v.visit(*this); }
 };
